Updated to support having multiple RCs
In the current state, only one RC at a time was supported. With major releases having a longer period in RC we may need to release minor/patch releases to the previous version while we have a major version RC out there.

This updates the tool to support any RC that did not have yet a non-rc release published by the same version.